requirements.txt
psycopg[c]>=3.2.6,<3.3
Django>=5.2,<5.3
django-filter>=24.0,<25.0
djangorestframework>=3.15.2,<3.16
django-vectortiles>=1.0.1,<1.1
djangorestframework-gis>=1.1,<1.2
django-silk>=5.3.2,<5.4
# django-hosts>=6.0,<6.1
django-cors-headers>=4.7.0,<4.8
drf-spectacular>=0.28.0,<0.29
django-lb-health-check>=1.0.1,<1.1
gunicorn>=23.0.0,<23.1.0

# Testing
pytest>=8.3.5,<8.4
pytest-django>=4.11.1,<4.12

# Others
geopandas>=1.0.1,<1.1
prophet>=1.1.6,<1.2
celery>=5.5.2,<5.6
flower>=2.0.1,<2.1
tqdm>=4.67.1,<4.68
ipython>=9.3.0,<9.4.0